move serde impls to separate module
